Disclamir
# Disclaimer


The end-user is solely responsible for utilizing the project and its resources. The developers bear no liability and cannot be held accountable for any harm or misuse resulting from the use of the project. It should be noted that certain actions may be unlawful, and this software cannot be employed to test individuals or organizations without obtaining written consent.